Game Mode integrated with Room  - implemented images being loaded with a set of 5  - send correct answer chat message  - added some getters, setters for these checks  - better log message when player disconnects  - fixed discrepancy when sending packet for loading scenes enjoy 🥳 we are nearing completion of the first release!  - fixed server tick rate bug

+++ Client/src/main/java/com/github/TeamRocketBalleBalle/Ricktionary/Client/Networking/PlayerNetworking.java
@@ -63,7 +63,7 @@ public class PlayerNetworking {
                         }
                     }
                     case PacketType.LOAD_SCENE -> {
-                        switch ((Byte) receivedPacket.getOrder().getValue()) {
+                        switch ((Integer) receivedPacket.getOrder().getValue()) {
                             case LoadScene.MATCHMAKING_SCENE -> test.sceneSwitch(
                                     "matchmakingscreen");
